feat: MapMenuのデフォルトセクションと型定義をmapMenuData.tsに追加
