+++ src/modules/map/ui/mapMenuData.ts
@@ -1,0 +1,47 @@
+export type MapMenuEntry = {
+  id: string;
+  label: string;
+  disabled?: boolean;
+  type?: "short" | "long";
+};
+
+export type MapMenuSection = MapMenuEntry & {
+  items?: MapMenuEntry[];
+};
+
+export const defaultMapMenuSections: MapMenuSection[] = [
+  {
+    id: "overall",
+    label: "全体",
+  },
+  {
+    id: "lecture-building",
+    label: "講義棟内",
+    items: [
+      { id: "lecture-building-1f", label: "1F", type: "short" },
+      { id: "lecture-building-2f", label: "2F", type: "short" },
+      { id: "lecture-building-3f", label: "3F", type: "short" },
+    ],
+  },
+  {
+    id: "outdoor-area",
+    label: "屋外エリア",
+    items: [
+      { id: "outdoor-overall", label: "屋外エリア全体" },
+      { id: "office-area", label: "事務棟エリア" },
+      { id: "library-area", label: "図書館エリア" },
+      { id: "electrical-area", label: "電気棟エリア" },
+      { id: "outdoor-stage-area", label: "屋外ステージエリア" },
+      { id: "mechanical-civil-area", label: "機械建設棟エリア" },
+      { id: "kitchen-car-area", label: "キッチンカーエリア" },
+    ],
+  },
+  {
+    id: "stamp-rally",
+    label: "スタンプラリー",
+  },
+  {
+    id: "mystery-solving",
+    label: "謎解き",
+  },
+];
